requirements.txt

requests
pandas
python-telegram-bot
nest-asyncio
openpyxl
